+++ front/components/agent_builder/triggers/AgentBuilderTriggersBlock.tsx
@@ -22,7 +22,7 @@ import { AgentBuilderSectionContainer } from "@app/components/agent_builder/Agen
 import { useSpacesContext } from "@app/components/agent_builder/SpacesContext";
 import { ScheduleEditionSheet } from "@app/components/agent_builder/triggers/ScheduleEditionSheet";
 import { TriggerCard } from "@app/components/agent_builder/triggers/TriggerCard";
-import { WebhookEdition } from "@app/components/agent_builder/triggers/WebhookEditionSheet";
+import { WebhookEdition } from "@app/components/agent_builder/triggers/webhook/WebhookEdition";
 import { getIcon } from "@app/components/resources/resources_icons";
 import { useSendNotification } from "@app/hooks/useNotification";
 import { useWebhookSourceViewsFromSpaces } from "@app/lib/swr/webhook_source";
